model: Make content and topic optional in update message requests.
The message content and topic for updating message should be not
passed to the request when they are not supposed to be edited.

While editing only the topic - if the content is also sent in the
request parameters, the server will return a 400 bad request. This
behaviour has not been updated in the REST API docs as of this commit,
but can be confirmed at
https://github.com/zulip/zulip/blob/7229088b6cd331ed8370fb11625346f1a5614c2b/zerver/views/messages.py#L1588

While editing only the content for stream messages - if topic is also
sent in the request parameters, the server will consider it as a
content+topic change. As, a consequence this unchanged topic edit
will show up in the edit history. Similar issue occurs when sending an
edit request when content is unchanged - for both private and stream
messages.

This commit makes content and topic optional parameters in
model.update_stream_message and model.update_stream_message.

The issue of actually not sending the topic/commit if it is unchanged
will be fixed in a later commit.

Tests amended.