model: Fetch and store subscribers for streams.
The optional flag include_subscribers in client.register is set to True
so that subscriber data for streams is returned. This data can be useful
for multiple features, such as ordering typeahead by subscribers, and
displaying number of subscribers in stream information.

This commit also appropriately handles the subscription event for
updating the stream subscribers.

Test added, renamed old subscription event handling test.