boxes: Prioritize user typeahead over group typeahead.
Show user-name suggestions first instead of groups, since they are used
more often. This is important when a user is using typeahead with an
empty query ( for example just '@' ) because in the future, when typeaheads
are ordered by the participants in the current narrow - having groups first
would not make this ordering apparent on empty queries.