boxes: Order mentions typeahead by recipients for private messages.
This commit orders typeahead for mentions such that participants of the
private message/group message are displayed first. WriteBox now stores
an additional attribute called subscriber_ids which stores the user_ids
of the subscribers in the conversation. The name 'subscribers' is chosen
instead of 'recipients' so that the same attribute can be used in the
future for stream messages.

Test added.